Initial commit as a separate package from irony-mode
This is an extraction of the auto-complete plugin that used to be
bundled with irony-mode. Making it a standalone package that depends on
irony-mode and auto-complete will the installation now that package
managers in Emacs are popular.

Also, since a plugin for company-mode is likely to see the light of day,
I assume people that use company-mode will prefer to avoid installing
auto-complete and vice versa.